Migrate file writers to FilesystemSink.
