+++ src/CMakeLists.txt
@@ -8,6 +8,9 @@ set(tgt acquire-driver-zarr)
 add_library(${tgt} MODULE
         common.hh
         common.cpp
+        writers/sink.hh
+        writers/filesystem.sink.hh
+        writers/filesystem.sink.cpp
         writers/writer.hh
         writers/writer.cpp
         writers/zarrv2.writer.hh
